Various updates to protect user data

+++ src/Treehouse.FitnessFrog/Controllers/EntriesController.cs
@@ -54,6 +54,8 @@ namespace Treehouse.FitnessFrog.Controllers
         {
             var viewModel = new EntriesAddViewModel();
 
+            viewModel.Entry.UserId = User.Identity.GetUserId();
+
             viewModel.Init(_activitiesRepository);
 
             return View(viewModel);
@@ -66,6 +68,9 @@ namespace Treehouse.FitnessFrog.Controllers
 
             if (ModelState.IsValid)
             {
+                var entry = viewModel.Entry;
+                entry.UserId = User.Identity.GetUserId();
+
                 _entriesRepository.Add(viewModel.Entry);
 
                 TempData["Message"] = "Your entry was successfully added!";
@@ -85,7 +90,9 @@ namespace Treehouse.FitnessFrog.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Entry entry = _entriesRepository.Get((int)id);
+            var userId = User.Identity.GetUserId();
+
+            Entry entry = _entriesRepository.Get((int)id, userId);
 
             if (entry == null)
             {
@@ -108,6 +115,16 @@ namespace Treehouse.FitnessFrog.Controllers
 
             if (ModelState.IsValid)
             {
+                var entry = viewModel.Entry;
+                var userId = User.Identity.GetUserId();
+
+                if (!_entriesRepository.EntryOwnedByUserId(entry.Id, userId))
+                {
+                    return HttpNotFound();
+                }
+
+                entry.UserId = userId;
+
                 _entriesRepository.Update(viewModel.Entry);
 
                 TempData["Message"] = "Your entry was successfully updated!";
@@ -127,7 +144,8 @@ namespace Treehouse.FitnessFrog.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Entry entry = _entriesRepository.Get((int)id);
+            var userId = User.Identity.GetUserId();
+            Entry entry = _entriesRepository.Get((int)id, userId);
 
             if (entry == null)
             {
@@ -140,6 +158,13 @@ namespace Treehouse.FitnessFrog.Controllers
         [HttpPost]
         public ActionResult Delete(int id)
         {
+            var userId = User.Identity.GetUserId();
+
+            if (!_entriesRepository.EntryOwnedByUserId(id, userId))
+            {
+                return HttpNotFound();
+            }
+
             _entriesRepository.Delete(id);
 
             TempData["Message"] = "Your entry was successfully deleted!";
